Add try-catch to examine the presence of a func.
Add try-cache to examine the presence of a new implemented function of
'gist-vim'.
It is required when that user does not use a latest gist-vim in them
environment and (somehow) accidentally specified
'g:unite_gist_recursively_loading' in them vimrc.

If the user specified the 'g:unite_gist_recursively_loading' and
the 'gist#list_recursively' is not found (determined from E117 exception), the
function displays the warnings and use 'gist#list' instead.

See PR #2 conversation for detail (in Japanese).